Add warning about newlines for Windows users
Currently we print our output with raw newline characters, which is fine
for Unix users but a bit unhelpful on Windows. I can't immediately
find the recommended way to output the host OS line separator character,
so for now just add a note to the readme to open the output file in
WordPad or something else more sensible than Notepad.

Signed-off-by: Peter Maydell <pmaydell@chiark.greenend.org.uk>